refactor: extract web app
